+++ src/y2021/d02/part2.rs
@@ -39,21 +39,21 @@
 
 use super::Direction;
 
-pub fn solve(input: &super::DirectionList) -> isize {
+pub fn solve(input: &super::DirectionList) -> usize {
     let mut aim = 0;
     let mut horizontal = 0;
     let mut depth = 0;
     for dir in input {
         match *dir {
             Direction::Forward(v) => {
-                horizontal += v;
-                depth += aim * v;
+                horizontal += v as isize;
+                depth += aim * v as isize;
             }
-            Direction::Up(v) => aim -= v,
-            Direction::Down(v) => aim += v,
+            Direction::Up(v) => aim -= v as isize,
+            Direction::Down(v) => aim += v as isize,
         }
     }
-    horizontal * depth
+    (horizontal * depth) as usize
 }
 
 #[cfg(test)]
